KSENROLL-6508	 - Support adding to the coloset one or more AOs with RDL/ADL information
git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-enroll/trunk@46873 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-enroll-ui/src/main/java/org/kuali/student/enrollment/class2/courseoffering/service/impl/ActivityOfferingMaintainableImpl.java
@@ -83,7 +83,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.Formatter;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -178,12 +177,12 @@ public class ActivityOfferingMaintainableImpl extends KSMaintainableImpl impleme
                         activity.getActivityOfferingInfo().setMaximumEnrollment(activityOfferingWrapper.getSharedMaxEnrollment());
                     }
 
+                    if (activityOfferingWrapper.isColocatedAO() && !activity.isAlreadyPersisted()){
+                        getScheduleHelper().deleteRequestedAndActualSchedules(activity.getActivityOfferingInfo());
+                    }
+
                     ActivityOfferingInfo updatedAO = getCourseOfferingService().updateActivityOffering(activity.getAoId(), activity.getActivityOfferingInfo(), createContextInfo());
                     activity.setActivityOfferingInfo(updatedAO);
-
-                    /*if (activityOfferingWrapper.isColocatedAO() && !activity.isAlreadyPersisted()){
-                        getScheduleHelper().deleteRequestedAndActualSchedules(activity.getActivityOfferingInfo());
-                    }*/
                 }
             } catch (Exception e) {
                 throw convertServiceExceptionsToUI(e);
